+++ cpp/l2-proxy/metrics_manager.cpp
@@ -27,3 +27,28 @@ prometheus::Histogram &MetricsManager::create_histogram(
       .Register(*registry)
       .Add({}, buckets);
 }
+
+prometheus::Family<prometheus::Counter> &
+MetricsManager::create_counter_family(
+    const std::shared_ptr<prometheus::Registry> &registry,
+    const std::string &name, const std::string &help) {
+  return prometheus::BuildCounter().Name(name).Help(help).Register(*registry);
+}
+
+prometheus::Family<prometheus::Gauge> &
+MetricsManager::create_gauge_family(
+    const std::shared_ptr<prometheus::Registry> &registry,
+    const std::string &name, const std::string &help) {
+  return prometheus::BuildGauge().Name(name).Help(help).Register(*registry);
+}
+
+prometheus::Family<prometheus::Histogram> &
+MetricsManager::create_histogram_family(
+    const std::shared_ptr<prometheus::Registry> &registry,
+    const std::string &name, const std::string &help,
+    const std::vector<double> &buckets) {
+  return prometheus::BuildHistogram()
+      .Name(name)
+      .Help(help)
+      .Register(*registry);
+}